requirements.txt
numpy==1.14.2
pandas==0.22.0
requests==2.22.0
Adafruit_GPIO==1.0.3
netifaces==0.10.9
paho_mqtt==1.5.0
pyserial==3.4
